Fix intermittent failure to purge s3 files.
- communityvew.py:
  + move s3 tree deletion to purge_images() such that it
    deletes s3 tree before attempting to delete incomming tree
  + improve exception handling in purge_images()
- webfs*.py: make rmtree(path) uniform across implementations
  so that no exception is raised if path does not exist
- TestCommunityView.py: improve purge test to put some files
  to be purged into s3 tree

+++ CommunityView/src/webfs_local.py
@@ -47,9 +47,11 @@ def path_isfile(filepath):
 def path_join(*args):
     return os.path.join(*args)
 
-# delete the tree of files rooted at path. Raise exception if deletion fails
+# delete the tree of files rooted at path. Raise exception if deletion fails.
+# However, do not raise an exception if path does not exist
 #
 def rmtree(path):
-    shutil.rmtree(path)
+    if os.path.isdir(path):
+        shutil.rmtree(path)
